requirements.txt
flask==3.0.3
mysql-connector-python==8.4.0
openai==1.30.5
python-dotenv==1.0.1
bcrypt==4.1.3